Lower s pointer to uintptr_t priot to ulong cast to support windows targets in src/include/symbol.h.

+++ src/include/symbol.h
@@ -58,7 +58,7 @@ inline int symbol::operator!=(symbol p) const
 
 inline unsigned long symbol::hash() const
 {
-  return (unsigned long)s;
+  return (unsigned long)(intptr_t)s;
 }
 
 inline const char *symbol::contents() const
